ajout securité pour changement hangar
Event de selection du spinner fire lors creation Fragment, du coup on enlever la caravane du hangar et on la remonter dedans, donc on perdait ses coordonnées dans le hangar... mise en place d'un test pour verifier qu'on change bien de hangar (test sur nom)

+++ Android/src/com/meeple/cloud/hivernage/view/clients/ClientInfoFragment.java
@@ -226,13 +226,15 @@ public class ClientInfoFragment extends Fragment implements TextWatcher{
 
     	String newhangarName = NewClientFragment.getAllHangarName()[hangarPos];
     	
-		Services.caravaneService.removeFromHangar(this.client.getCaravane());
-		
-		Hangar newHangar = Services.hangarService.findHangarByName(newhangarName);
-		
-		if (newHangar != null) {
-			Services.caravaneService.putInHangar(this.client.getCaravane(), new EmplacementHangar(0, 0, 0.0d, newHangar));
-		}
+    	if (!newhangarName.equals(this.client.getCaravane().getEmplacementHangar().getHangar().getNom())) {
+    	
+			Hangar newHangar = Services.hangarService.findHangarByName(newhangarName);
+			
+			if (newHangar != null) {
+				Services.caravaneService.removeFromHangar(this.client.getCaravane());
+				Services.caravaneService.putInHangar(this.client.getCaravane(), new EmplacementHangar(0, 0, 0.0d, newHangar));
+			}
+    	}
 	}
     
     protected void changeGabarit(int gabaritPos) {
